.......
I was busy working on C# for a while for that internship, but, it did not work out. I like C#, but, I have to continue to work on JS, so I can help with interns. I just finished some work on cycle10rfp. It's functional. I may touch it up a bit. but for now I want to get back into diving deep into JS